+++ src/test/java/data/tree/ConstructBinaryTreeFromPreorderAndInorderTraversalTest.java
@@ -25,11 +25,11 @@ public class ConstructBinaryTreeFromPreorderAndInorderTraversalTest {
         //   9  20
         //     /  \
         //    15   7
-        assertEquals(3, root.val.intValue());
-        assertEquals(9, root.left.val.intValue());
-        assertEquals(20, root.right.val.intValue());
-        assertEquals(15, root.right.left.val.intValue());
-        assertEquals(7, root.right.right.val.intValue());
+        assertEquals(3, root.val);
+        assertEquals(9, root.left.val);
+        assertEquals(20, root.right.val);
+        assertEquals(15, root.right.left.val);
+        assertEquals(7, root.right.right.val);
     }
 
 }
